refactor: modulariza as classes do projeto
